refactor: use features pattern
