+++ src/app/(app)/(chats)/index.tsx
@@ -1,6 +1,6 @@
 import { Text } from "react-native";
 
-import { Container } from "@/styles/app/chats/styles";
+import { Container } from "@/features/app/chats/styles";
 
 export default function ChatsPage() {
   return (
