+++ src/app/(app)/(create)/index.tsx
@@ -1,6 +1,6 @@
 import { Text } from "react-native";
 
-import { Container } from "@/styles/app/create/styles";
+import { Container } from "@/features/app/create/styles";
 
 export default function CreatePage() {
   return (
