+++ src/app/(app)/(memes)/(trending)/index.tsx
@@ -1,13 +1,13 @@
 import React, { useEffect, useState } from "react";
 import { FlatList, useWindowDimensions } from "react-native";
 
+import { Meme } from "@/components/meme";
 import {
   fetchTrendingMemes,
   type TrendingMeme,
-} from "@/api/fetch-trending-memes";
-import { Meme } from "@/components/meme";
+} from "@/features/app/memes/api/fetch-trending-memes";
 
-import { Container } from "@/styles/app/memes/trending.styles";
+import { Container } from "@/features/app/memes/trending.styles";
 
 type FeedItem = TrendingMeme;
