+++ src/app/(auth)/_layout.tsx
@@ -1,7 +1,7 @@
 import { Slot } from "expo-router";
 import React from "react";
 
-import { FormContainer, LayoutContainer } from "@/styles/auth/styles";
+import { FormContainer, LayoutContainer } from "@/features/auth/styles/layout";
 
 import authBackground from "@/assets/auth-background.jpg";
