+++ src/features/auth/styles/forgot-password.ts
