+++ src/features/auth/styles/register.ts
